requirements.txt
requests
shapely
mysql-connector-python
